[Add] Event(객체 지연 생성 및 제거, 제거 시 객체 간 충돌 관련 후처리)

+++ 2023_Win32ApiStudy/AssetManager.cpp
@@ -31,7 +31,7 @@ void CAssetManager::Init()
 	GetCurrentDirectory(255, m_assetPath);
 
 	// 상위 폴더 경로를 구한다.
-	for (int i = wcslen(m_assetPath) - 1; i >= 0; --i)
+	for (int i = (int)wcslen(m_assetPath) - 1; i >= 0; --i)
 	{
 		if (m_assetPath[i] == '\\')
 		{
